requirements.txt
# Scashyyyy Personal Assistant
flask>=3.0.0
groq>=0.4.0
python-dotenv>=1.0.0
duckduckgo-search>=4.0.0  # Free web search
gunicorn>=21.0.0  # Production server
